% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navDownloadButton.R
\name{navDownloadButtonDependencies}
\alias{navDownloadButtonDependencies}
\title{Register dependencies for the navbar download button}
\usage{
navDownloadButtonDependencies()
}
\description{
Register dependencies for the navbar download button
}
\keyword{internal}
