% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triStateCheckbox.R
\name{triStateCheckboxDependencies}
\alias{triStateCheckboxDependencies}
\title{Register dependencies for tri-state checkbox}
\usage{
triStateCheckboxDependencies()
}
\description{
Register dependencies for tri-state checkbox
}
\keyword{internal}
