% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triStateCheckbox.R
\name{updateTriStateCheckboxInput}
\alias{updateTriStateCheckboxInput}
\title{Update a tri-state checkbox input}
\usage{
updateTriStateCheckboxInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL
)
}
\arguments{
\item{session}{The Shiny session object (default is getDefaultReactiveDomain())}

\item{inputId}{The ID of the input object to update}

\item{label}{The new label (optional)}

\item{value}{The new value (optional): -1 (exclude), 0 (neutral), or 1 (include)}
}
\description{
Change the value and/or label of a tri-state checkbox input on the client.
}
\examples{
\dontrun{
# In server function
updateTriStateCheckboxInput(session, "checkbox", label = "New Label", value = 1)
}

}
