% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoRadioButtons.R
\name{generateInfoPretty}
\alias{generateInfoPretty}
\title{Internal function to generate pretty radio options with info bubbles}
\usage{
generateInfoPretty(
  inputId,
  selected,
  inline,
  type = "checkbox",
  choiceNames,
  choiceValues,
  descriptions = NULL,
  status = "primary",
  shape = "square",
  outline = FALSE,
  fill = FALSE,
  thick = FALSE,
  animation = NULL,
  icon = NULL,
  plain = FALSE,
  bigger = FALSE
)
}
\description{
Internal function to generate pretty radio options with info bubbles
}
