% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triStateCheckbox.R
\name{triStateCheckboxInput}
\alias{triStateCheckboxInput}
\title{Create a tri-state checkbox input control}
\usage{
triStateCheckboxInput(inputId, label, value = 0)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value: -1 (exclude), 0 (neutral), or 1 (include)}
}
\value{
A custom Shiny input control that can be added to a UI definition
}
\description{
This function creates a tri-state checkbox input that lets users cycle
through three states: Include (1), Exclude (-1), and Neither/Neutral (0).
The checkbox visually indicates the state with different icons and colors.
}
\examples{
\dontrun{
library(shiny)
ui <- fluidPage(
  triStateCheckboxInput("checkbox", "Select option", 0)
)
server <- function(input, output) {
  observeEvent(input$checkbox, {
    print(getTriStateLabel(input$checkbox))
  })
}
shinyApp(ui, server)
}

}
